Organize hooks file
