+++ src/hooks.server.ts
@@ -3,49 +3,17 @@ import { DEMO_USER } from "$lib";
 import { getCurrentFormattedDateTime } from "$lib/server";
 import Labline from "$lib/server/api/Labline";
 import * as auth from "$lib/server/auth";
-import { User } from "$lib/server/db/repository";
+import { onServerStart } from "$lib/server/init";
 import { Security } from "$lib/server/Security";
-import { error, type Handle } from "@sveltejs/kit";
+import { error, type Handle, type ServerInit } from "@sveltejs/kit";
 import { sequence } from "@sveltejs/kit/hooks";
 
-const BYPASS_ACCOUNT_REQUIREMENT = env.BYPASS_ACCOUNT_REQUIREMENT === "true";
-const REQUIRED_ENV_VARIABLES = [
-	"DATABASE_URL",
-	"OIDC_DISCOVERY_ENDPOINT",
-	"OIDC_CLIENT_ID",
-	"OIDC_CLIENT_SECRET",
-	"ABSOLUTE_DIR_PATH",
-];
-
-const missingVars = [];
-for (const envVar of REQUIRED_ENV_VARIABLES) {
-	if (!env[envVar]) {
-		missingVars.push(envVar);
-	}
-}
-if (missingVars.length !== 0) {
-	console.log(`${missingVars.join(", ")} must be set`);
-	process.exit(1);
-}
-
-if (env.CREATE_ACCOUNT) {
-	try {
-		console.log("Creating user from CREATE_ACCOUNT...");
-		const newUser = await User.createUser(
-			env.CREATE_ACCOUNT,
-			"Initial User",
-			["admin", "superadmin"],
-			true,
-		);
-		console.log(`Created user ${newUser.username}`);
-	} catch (e) {
-		if (e instanceof Error) {
-			console.log("Error creating initial account: ", e.message);
-		}
-	}
-}
+export const init: ServerInit = async () => {
+	await onServerStart();
+};
 
 const originalHandle: Handle = async ({ event, resolve }) => {
+	const BYPASS_ACCOUNT_REQUIREMENT = env.BYPASS_ACCOUNT_REQUIREMENT === "true";
 	if (BYPASS_ACCOUNT_REQUIREMENT && !event.locals.user) {
 		event.locals.user = DEMO_USER;
 	}
